requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-openai>=0.0.8
mem0ai>=0.1.0
python-dotenv>=1.0.0
openai>=1.0.0
chromadb>=0.4.0
tiktoken>=0.5.0 